// Next.js/Lez32/models/requests/user-requests.ts
export interface GetFPUsersRequest {
    page?: number;
    limit?: number;
    filter?: string;
    isActive?: string;
}

export interface SaveUserRequest {
    id?: string;
    username: string;
    email: string;
    password?: string;
    firstName: string;
    middleName?: string;
    lastName: string;
    role?: number;
    isActive?: boolean;
}